+++ swingexplorer-core/src/main/java/com/example/CheckThreadViolationRepaintManager.java
@@ -17,7 +17,7 @@
  *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *   
  */
-package sample;
+package com.example;
 
 
 import java.lang.ref.WeakReference;
