+++ swingexplorer-core/src/main/java/com/example/FrmPerson.form
@@ -239,7 +239,7 @@
         <Property name="font" type="java.awt.Font" editor="org.netbeans.beaninfo.editors.FontEditor">
           <Font name="Tahoma" size="12" style="1"/>
         </Property>
-        <Property name="text" type="java.lang.String" value="This is sample Swing application to demonstrate Swing Explorer"/>
+        <Property name="text" type="java.lang.String" value="This is a sample Swing application to demonstrate Swing Explorer"/>
       </Properties>
     </Component>
     <Component class="javax.swing.JButton" name="btnModalDialog">
