+++ swingexplorer-core/src/main/java/com/example/PersonEditorWithSwingExplorer.java
@@ -1,0 +1,7 @@
+package com.example;
+
+public class PersonEditorWithSwingExplorer {
+    public static void main(String args[]) {
+        org.swingexplorer.Launcher.main(new String[] {"com.example.PersonEditor"});
+    }
+}
